Simplify event counter local-core interface
