+++ litebox/src/event/counter.rs
@@ -1,22 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+use alloc::sync::Arc;
+
+pub use litebox_broker_protocol::EventConsumeMode as EventCounterReadMode;
+use litebox_broker_protocol::{
+    AddEventRequest, ConsumeEventRequest, ConsumeEventResponse, CoreRequest, CoreResponse,
+    CreateEventRequest, EventRequest, EventResponse, ObjectHandle, ReadinessState,
+    WaitEventRequest, WaitOutcome,
+};
+
 use crate::{
-    broker::event::BrokerEventCounter,
-    event::{Events, IOPollable, observer::Observer, polling::TryOpError, wait::WaitContext},
+    LiteBox,
+    broker::{
+        BrokerControl,
+        error::{BrokerObjectError, map_broker_object_result},
+    },
+    event::{
+        Events, IOPollable, observer::Observer, polling::Pollee, polling::TryOpError,
+        wait::WaitContext,
+    },
     platform::TimeProvider,
     sync::RawSyncPrimitivesProvider,
 };
-
-/// How an event-counter read consumes the current counter value.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum EventCounterReadMode {
-    /// Consume the full counter value.
-    All,
-    /// Consume one counter credit.
-    One,
-}
 
 /// Errors returned by local-core event counters.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -36,60 +42,117 @@ pub enum EventCounterError {
     Unavailable,
 }
 
-enum EventCounterBackend<Platform: RawSyncPrimitivesProvider + TimeProvider> {
-    Broker(BrokerEventCounter<Platform>),
-}
-
 /// A local-core event counter object.
-///
-/// The object API is independent of the current backing implementation. The
-/// split-broker POC uses a broker-backed counter, while future migrations can add
-/// other backends without changing shim-facing call sites.
 pub struct EventCounter<Platform: RawSyncPrimitivesProvider + TimeProvider> {
-    backend: EventCounterBackend<Platform>,
+    broker: Arc<dyn BrokerControl>,
+    handle: ObjectHandle,
+    pollee: Pollee<Platform>,
+    blocking_operations_supported: bool,
 }
 
 impl<Platform> EventCounter<Platform>
 where
     Platform: RawSyncPrimitivesProvider + TimeProvider,
 {
-    pub(crate) fn from_broker(event: BrokerEventCounter<Platform>) -> Self {
-        Self {
-            backend: EventCounterBackend::Broker(event),
-        }
+    /// Creates a local-core event counter.
+    pub fn new(litebox: &LiteBox<Platform>, initial_count: u64) -> Result<Self, EventCounterError> {
+        let Some(broker) = litebox.broker_control() else {
+            return Err(EventCounterError::Unavailable);
+        };
+        let response = broker
+            .request(CoreRequest::Event(EventRequest::Create(
+                CreateEventRequest::new(initial_count),
+            )))
+            .map_err(BrokerObjectError::from)
+            .and_then(event_response_from_core)
+            .map_err(EventCounterError::from)?;
+        let EventResponse::Create(response) = response else {
+            return Err(BrokerObjectError::UnexpectedResponse.into());
+        };
+        Ok(Self {
+            broker,
+            handle: response.handle,
+            pollee: Pollee::new(),
+            blocking_operations_supported: false,
+        })
     }
 
     /// Returns whether blocking reads and writes are supported.
     pub fn supports_blocking_operations(&self) -> bool {
-        match &self.backend {
-            EventCounterBackend::Broker(_) => {
-                BrokerEventCounter::<Platform>::supports_blocking_operations()
-            }
-        }
+        self.blocking_operations_supported
     }
 
     /// Reads the event counter.
     pub fn read(
         &self,
-        cx: &WaitContext<'_, Platform>,
-        nonblock: bool,
+        _cx: &WaitContext<'_, Platform>,
+        _nonblock: bool,
         mode: EventCounterReadMode,
     ) -> Result<u64, TryOpError<EventCounterError>> {
-        match &self.backend {
-            EventCounterBackend::Broker(event) => event.read(cx, nonblock, mode),
+        let response = map_broker_object_result(self.consume(mode))?;
+        if response.readiness.write_ready {
+            self.pollee.notify_observers(Events::OUT);
         }
+        Ok(response.value)
     }
 
     /// Writes readiness credits to the event counter.
     pub fn write(
         &self,
-        cx: &WaitContext<'_, Platform>,
-        nonblock: bool,
+        _cx: &WaitContext<'_, Platform>,
+        _nonblock: bool,
         value: u64,
     ) -> Result<usize, TryOpError<EventCounterError>> {
-        match &self.backend {
-            EventCounterBackend::Broker(event) => event.write(cx, nonblock, value),
+        if value == u64::MAX {
+            return Err(TryOpError::Other(EventCounterError::InvalidInput));
         }
+        let readiness = map_broker_object_result(self.add(value))?;
+        if value != 0 && readiness.read_ready {
+            self.pollee.notify_observers(Events::IN);
+        }
+        Ok(core::mem::size_of::<u64>())
+    }
+
+    fn consume(
+        &self,
+        mode: EventCounterReadMode,
+    ) -> Result<ConsumeEventResponse, BrokerObjectError> {
+        let response = self.request_event(EventRequest::Consume(ConsumeEventRequest::new(
+            self.handle,
+            mode,
+        )))?;
+        let EventResponse::Consume(response) = response else {
+            return Err(BrokerObjectError::UnexpectedResponse);
+        };
+        Ok(response)
+    }
+
+    fn add(&self, value: u64) -> Result<ReadinessState, BrokerObjectError> {
+        let response =
+            self.request_event(EventRequest::Add(AddEventRequest::new(self.handle, value)))?;
+        let EventResponse::Add(response) = response else {
+            return Err(BrokerObjectError::UnexpectedResponse);
+        };
+        Ok(response.readiness)
+    }
+
+    fn readiness_state(&self) -> Result<ReadinessState, BrokerObjectError> {
+        let response =
+            self.request_event(EventRequest::Wait(WaitEventRequest::new(self.handle)))?;
+        let EventResponse::Wait(response) = response else {
+            return Err(BrokerObjectError::UnexpectedResponse);
+        };
+        Ok(match response.outcome {
+            WaitOutcome::Ready(readiness) | WaitOutcome::WouldBlock(readiness) => readiness,
+            _ => return Err(BrokerObjectError::UnexpectedResponse),
+        })
+    }
+
+    fn request_event(&self, request: EventRequest) -> Result<EventResponse, BrokerObjectError> {
+        self.broker
+            .request(CoreRequest::Event(request))
+            .map_err(BrokerObjectError::from)
+            .and_then(event_response_from_core)
     }
 }
 
@@ -98,14 +161,219 @@ where
     Platform: RawSyncPrimitivesProvider + TimeProvider,
 {
     fn register_observer(&self, observer: alloc::sync::Weak<dyn Observer<Events>>, mask: Events) {
-        match &self.backend {
-            EventCounterBackend::Broker(event) => event.register_observer(observer, mask),
-        }
+        self.pollee.register_observer(observer, mask);
     }
 
     fn check_io_events(&self) -> Events {
-        match &self.backend {
-            EventCounterBackend::Broker(event) => event.check_io_events(),
+        let Ok(readiness) = self.readiness_state() else {
+            return Events::empty();
+        };
+        let mut events = Events::empty();
+        if readiness.read_ready {
+            events |= Events::IN;
         }
+        if readiness.write_ready {
+            events |= Events::OUT;
+        }
+        events
+    }
+}
+
+fn event_response_from_core(response: CoreResponse) -> Result<EventResponse, BrokerObjectError> {
+    match response {
+        CoreResponse::Event(response) => Ok(response),
+        _ => Err(BrokerObjectError::UnexpectedResponse),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    extern crate std;
+
+    use alloc::sync::Arc;
+    use core::sync::atomic::{AtomicUsize, Ordering};
+    use std::{collections::VecDeque, sync::Mutex};
+
+    use litebox_broker_protocol::{
+        AddEventResponse, ConsumeEventResponse, CoreRequest, CoreResponse, CreateEventResponse,
+        EventResponse, ObjectHandle, ObjectReferenceGeneration, ObjectReferenceId, ReadinessState,
+        WaitEventResponse,
+    };
+
+    use super::*;
+    use crate::{
+        BrokerControlError,
+        event::{Events, IOPollable, observer::Observer, wait::WaitState},
+        platform::mock::MockPlatform,
+    };
+
+    struct MockBrokerControl {
+        responses: Mutex<VecDeque<CoreResponse>>,
+    }
+
+    impl MockBrokerControl {
+        fn new(responses: impl IntoIterator<Item = CoreResponse>) -> Self {
+            Self {
+                responses: Mutex::new(responses.into_iter().collect()),
+            }
+        }
+    }
+
+    impl BrokerControl for MockBrokerControl {
+        fn request(
+            &self,
+            _request: CoreRequest,
+        ) -> core::result::Result<CoreResponse, BrokerControlError> {
+            self.responses
+                .lock()
+                .unwrap()
+                .pop_front()
+                .ok_or(BrokerControlError::UnexpectedResponse)
+        }
+    }
+
+    struct CountingObserver {
+        notifications: AtomicUsize,
+    }
+
+    impl CountingObserver {
+        fn new() -> Self {
+            Self {
+                notifications: AtomicUsize::new(0),
+            }
+        }
+
+        fn notifications(&self) -> usize {
+            self.notifications.load(Ordering::Relaxed)
+        }
+    }
+
+    impl Observer<Events> for CountingObserver {
+        fn on_events(&self, _events: &Events) {
+            self.notifications.fetch_add(1, Ordering::Relaxed);
+        }
+    }
+
+    fn sample_handle() -> ObjectHandle {
+        ObjectHandle::new(ObjectReferenceId::new(1), ObjectReferenceGeneration::new(1))
+    }
+
+    fn create_response() -> CoreResponse {
+        CoreResponse::Event(EventResponse::Create(CreateEventResponse::new(
+            sample_handle(),
+        )))
+    }
+
+    fn add_response(read_ready: bool, write_ready: bool) -> CoreResponse {
+        CoreResponse::Event(EventResponse::Add(AddEventResponse::new(
+            ReadinessState::new(read_ready, write_ready, 1),
+        )))
+    }
+
+    fn consume_response(value: u64, read_ready: bool, write_ready: bool) -> CoreResponse {
+        CoreResponse::Event(EventResponse::Consume(ConsumeEventResponse::new(
+            value,
+            ReadinessState::new(read_ready, write_ready, 1),
+        )))
+    }
+
+    fn observer() -> (Arc<CountingObserver>, Arc<dyn Observer<Events>>) {
+        let observer = Arc::new(CountingObserver::new());
+        let observer_dyn: Arc<dyn Observer<Events>> = observer.clone();
+        (observer, observer_dyn)
+    }
+
+    fn litebox(responses: impl IntoIterator<Item = CoreResponse>) -> LiteBox<MockPlatform> {
+        LiteBox::new_with_broker_control(
+            MockPlatform::new(),
+            Arc::new(MockBrokerControl::new(responses)),
+        )
+    }
+
+    #[test]
+    fn create_maps_unexpected_response_shape() {
+        let litebox = litebox([add_response(false, true)]);
+        let result = EventCounter::<MockPlatform>::new(&litebox, 0);
+
+        assert!(matches!(result, Err(EventCounterError::UnexpectedResponse)));
+    }
+
+    #[test]
+    fn zero_write_does_not_notify_read_observers() {
+        let litebox = litebox([create_response(), add_response(false, true)]);
+        let event = EventCounter::new(&litebox, 0).unwrap();
+        let (observer, observer_dyn) = observer();
+        event.register_observer(Arc::downgrade(&observer_dyn), Events::IN);
+
+        let wait = WaitState::new(MockPlatform::new());
+        assert_eq!(event.write(&wait.context(), true, 0).unwrap(), 8);
+
+        assert_eq!(observer.notifications(), 0);
+    }
+
+    #[test]
+    fn nonzero_write_notifies_read_observers_when_broker_reports_ready() {
+        let litebox = litebox([create_response(), add_response(true, true)]);
+        let event = EventCounter::new(&litebox, 0).unwrap();
+        let (observer, observer_dyn) = observer();
+        event.register_observer(Arc::downgrade(&observer_dyn), Events::IN);
+
+        let wait = WaitState::new(MockPlatform::new());
+        assert_eq!(event.write(&wait.context(), true, 1).unwrap(), 8);
+
+        assert_eq!(observer.notifications(), 1);
+    }
+
+    #[test]
+    fn read_does_not_notify_write_observers_without_broker_write_readiness() {
+        let litebox = litebox([create_response(), consume_response(1, false, false)]);
+        let event = EventCounter::new(&litebox, 0).unwrap();
+        let (observer, observer_dyn) = observer();
+        event.register_observer(Arc::downgrade(&observer_dyn), Events::OUT);
+
+        let wait = WaitState::new(MockPlatform::new());
+        assert_eq!(
+            event
+                .read(&wait.context(), true, EventCounterReadMode::All)
+                .unwrap(),
+            1
+        );
+
+        assert_eq!(observer.notifications(), 0);
+    }
+
+    #[test]
+    fn read_notifies_write_observers_when_broker_reports_write_ready() {
+        let litebox = litebox([create_response(), consume_response(1, false, true)]);
+        let event = EventCounter::new(&litebox, 0).unwrap();
+        let (observer, observer_dyn) = observer();
+        event.register_observer(Arc::downgrade(&observer_dyn), Events::OUT);
+
+        let wait = WaitState::new(MockPlatform::new());
+        assert_eq!(
+            event
+                .read(&wait.context(), true, EventCounterReadMode::All)
+                .unwrap(),
+            1
+        );
+
+        assert_eq!(observer.notifications(), 1);
+    }
+
+    #[test]
+    fn poll_uses_broker_read_and_write_readiness() {
+        let litebox = litebox([
+            create_response(),
+            CoreResponse::Event(EventResponse::Wait(WaitEventResponse::new(
+                WaitOutcome::Ready(ReadinessState::new(true, false, 1)),
+            ))),
+            CoreResponse::Event(EventResponse::Wait(WaitEventResponse::new(
+                WaitOutcome::WouldBlock(ReadinessState::new(false, true, 2)),
+            ))),
+        ]);
+        let event = EventCounter::new(&litebox, 0).unwrap();
+
+        assert_eq!(event.check_io_events(), Events::IN);
+        assert_eq!(event.check_io_events(), Events::OUT);
     }
 }
